refactor(solidFromSlices): moved solidFromSlices to a seperate module
